Fix compilation issues and add library crate
